moved component to folder named after itself + Fiz review
