+++ webapp/src/shared/components/index.ts
@@ -51,4 +51,4 @@ export { default as Game } from './Game/Game';
 
 export { default as GameTraining } from './Game/GameTraining';
 
-export { default as CustomConfirmAlert } from './ConfirmAlert/CustomConfirmAlert';
+export { default as CustomConfirmAlert } from './CustomConfirmAlert/CustomConfirmAlert';
